move extractHeadings to non-client utils file to fix docs + blog table of contents
