+++ docs/app/(site)/blog/[post]/page-client.tsx
@@ -6,11 +6,12 @@ import Link from 'next/link'
 import { useParams } from 'next/navigation'
 import { parse, format } from 'date-fns'
 
-import { extractHeadings, Markdoc } from '../../../../components/Markdoc'
+import {  Markdoc } from '../../../../components/Markdoc'
 import { BlogPage } from '../../../../components/Page'
 import { Heading } from '../../../../components/docs/Heading'
 import { Type } from '../../../../components/primitives/Type'
 import { type BlogPost } from './page'
+import { extractHeadings } from '../../../../markdoc/headings'
 
 export default function Page ({ post }: { post: BlogPost }) {
   const params = useParams()
